Improve the file command, implement query of specific pages, rename file to wikimedia.js
